+++ src/videopython/editing/__init__.py
@@ -1,0 +1,9 @@
+from .multicam import CutPoint, MultiCamEdit
+from .video_edit import SegmentConfig, VideoEdit
+
+__all__ = [
+    "VideoEdit",
+    "SegmentConfig",
+    "MultiCamEdit",
+    "CutPoint",
+]